+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.3
+
+IMPROVEMENTS:
+* Added CHAP, Personality, and Preferred Array properties to Host Struct.
+* Added params to ListHosts and GetHost to be able to query for above properties.
+
+## 0.1.2
+
+Because I messed up my tags
+
 ## 0.1.1 (December 30, 2018)
 
 Fixes: